UI: Move the plugins context to app level
Issue #3418
PR #3629
