+++ monkey/monkey_island/cc/ui/package.json
@@ -99,13 +99,13 @@
     "react-copy-to-clipboard": "^5.1.0",
     "react-desktop-notification": "^1.0.9",
     "react-dom": "^18.2.0",
-    "react-vis-graph-wrapper": "^0.1.3",
     "react-json-tree": "^0.18.0",
     "react-markdown": "^8.0.7",
-    "react-router-dom": "^6.12.1",
+    "react-router-dom": "^6.15.0",
     "react-spinners": "^0.13.8",
     "react-timer-hook": "^3.0.6",
     "react-tsparticles": "^2.10.1",
+    "react-vis-graph-wrapper": "^0.1.3",
     "remark-breaks": "^3.0.3",
     "request": "^2.88.2",
     "semver": "^7.5.2",
