+++ monkey/monkey_island/cc/ui/src/components/Main.tsx
@@ -19,9 +19,10 @@ import LoadingScreen from './ui-components/LoadingScreen';
 import SidebarLayoutComponent from "./layouts/SidebarLayoutComponent";
 import {CompletedSteps} from "./side-menu/CompletedSteps";
 import Timeout = NodeJS.Timeout;
-import { doesAnyAgentExist, didAllAgentsShutdown } from './utils/ServerUtils';
+import {doesAnyAgentExist, didAllAgentsShutdown} from './utils/ServerUtils';
 import LogoutPage from './pages/LogoutPage';
 import MarketplacePage from './pages/MarketplacePage';
+import PluginsProvider from './contexts/plugins/PluginsProvider';
 
 let notificationIcon = require('../images/notification-logo-512x512.png');
 
@@ -42,20 +43,55 @@ export const IslandRoutes = {
   PluginsPage: '/marketplace',
 }
 
-export function isReportRoute(route){
+export function isReportRoute(route) {
   return route.startsWith(IslandRoutes.Report);
 }
 
 class AppComponent extends AuthComponent {
   private interval: Timeout;
+  private pluginsFetched: boolean;
 
   constructor(props) {
     super(props);
     this.state = {
       completedSteps: new CompletedSteps(false),
+      availablePlugins: [],
+      installedPlugins: [],
+      numberOfPluginsThatRequiresUpdate: 0
     };
     this.interval = undefined;
+    this.pluginsFetched = false;
   }
+
+  setAvailablePlugins = (plugins) => {
+    this.setState({
+        availablePlugins: plugins
+    });
+  }
+
+  setInstalledPlugins = (plugins) => {
+    this.setState({
+        installedPlugins: plugins
+    });
+  }
+
+  // TODO: move to plugins utils?
+  refreshAvailablePlugins = (forceRefresh: boolean = false) => {
+    let url = '/api/agent-plugins/available/index';
+    if (forceRefresh) {
+      url += '?force_refresh=true';
+    }
+    this.authFetch(url, {}, true).then(res => res.json()).then(plugins => {
+      this.setAvailablePlugins(plugins?.plugins || []);
+    });
+  };
+
+  // TODO: move to plugins utils?
+  refreshInstalledPlugins = () => {
+    this.authFetch('/api/agent-plugins/installed/manifests', {}, true).then(res => res.json()).then(plugins => {
+      this.setInstalledPlugins(plugins);
+    });
+  };
 
   updateStatus = () => {
     if (this.state.isLoggedIn === false) {
@@ -80,16 +116,24 @@ class AppComponent extends AuthComponent {
     }
 
     if (res) {
+      // TODO: get plugins only one time
+      // TODO: Use real endpoint
+      if(!this.pluginsFetched) {
+        this.refreshAvailablePlugins();
+        this.refreshInstalledPlugins();
+          this.pluginsFetched = true;
+      }
+
       // update status: report generation
       this.authFetch('/api/report-generation-status', {}, false)
         .then(res => res.json())
         .then(res => {
           this.setState({
             completedSteps: new CompletedSteps(
-                                  this.state.completedSteps.runMonkey,
-                                  this.state.completedSteps.infectionDone,
-                                  res.report_done
-                                )
+              this.state.completedSteps.runMonkey,
+              this.state.completedSteps.infectionDone,
+              res.report_done
+            )
           });
         })
 
@@ -97,26 +141,25 @@ class AppComponent extends AuthComponent {
       doesAnyAgentExist(false).then(anyAgentExists => {
         this.setState({
           completedSteps: new CompletedSteps(
-                                anyAgentExists,
-                                this.state.completedSteps.infectionDone,
-                                this.state.completedSteps.reportDone
-                              )
+            anyAgentExists,
+            this.state.completedSteps.infectionDone,
+            this.state.completedSteps.reportDone
+          )
         });
       });
 
       // update status: if infection (running and shutting down of all agents) finished
       didAllAgentsShutdown(false).then(allAgentsShutdown => {
         let infectionDone = this.state.completedSteps.runMonkey && allAgentsShutdown;
-        if(this.state.completedSteps.infectionDone === false
-          && infectionDone){
+        if (this.state.completedSteps.infectionDone === false && infectionDone) {
           this.showInfectionDoneNotification();
         }
         this.setState({
           completedSteps: new CompletedSteps(
-                                this.state.completedSteps.runMonkey,
-                                infectionDone,
-                                this.state.completedSteps.reportDone
-                              )
+            this.state.completedSteps.runMonkey,
+            infectionDone,
+            this.state.completedSteps.reportDone
+          )
         });
       });
     }
@@ -165,12 +208,26 @@ class AppComponent extends AuthComponent {
   }
 
   render() {
+    const {availablePlugins, installedPlugins, numberOfPluginsThatRequiresUpdate} = this.state;
+    const pluginsContextValue = {
+      availablePlugins,
+      installedPlugins,
+      numberOfPluginsThatRequiresUpdate,
+      setAvailablePlugins: this.setAvailablePlugins,
+      setInstalledPlugins: this.setInstalledPlugins,
+      refreshAvailablePlugins: this.refreshAvailablePlugins,
+      refreshInstalledPlugins: this.refreshInstalledPlugins
+    }
 
-    let defaultSideNavProps = {completedSteps: this.state.completedSteps,
-                               onStatusChange: this.updateStatus,
-                               defaultReport: this.getDefaultReport(),
-                               onLogout: () => {this.auth.logout()
-                                 .then(() => this.updateStatus())}};
+    let defaultSideNavProps = {
+      completedSteps: this.state.completedSteps,
+      onStatusChange: this.updateStatus,
+      defaultReport: this.getDefaultReport(),
+      onLogout: () => {
+        this.auth.logout()
+          .then(() => this.updateStatus())
+      }
+    };
 
     return (
       <Router>
@@ -178,29 +235,32 @@ class AppComponent extends AuthComponent {
           <Routes>
             <Route path={IslandRoutes.LoginPage} element={<LoginPageComponent onStatusChange={this.updateStatus}/>}/>
             <Route path={IslandRoutes.Logout} element={<LogoutPage onStatusChange={this.updateStatus}/>}/>
-            <Route path={IslandRoutes.RegisterPage} element={<RegisterPageComponent onStatusChange={this.updateStatus}/>}/>
-            {this.renderRoute(IslandRoutes.GettingStartedPage,
-              <SidebarLayoutComponent component={GettingStartedPage} {...defaultSideNavProps}/>)}
-            {this.renderRoute(IslandRoutes.ConfigurePage,
-              <SidebarLayoutComponent component={ConfigurePage} {...defaultSideNavProps}/>)}
-            {this.renderRoute(IslandRoutes.RunMonkeyPage,
-              <SidebarLayoutComponent component={RunMonkeyPage} {...defaultSideNavProps}/>)}
-            {this.renderRoute(IslandRoutes.MapPage,
-              <SidebarLayoutComponent component={MapPageWrapper} {...defaultSideNavProps}/>)}
-            {this.renderRoute(IslandRoutes.EventPage,
-              <SidebarLayoutComponent component={EventPage} {...defaultSideNavProps}/>)}
-            {this.renderRoute(IslandRoutes.Marketplace,
-              <SidebarLayoutComponent component={MarketplacePage} {...defaultSideNavProps}/>)}
-            {this.redirectToReport()}
-            {this.renderRoute(IslandRoutes.SecurityReport,
-              <SidebarLayoutComponent component={ReportPage}
-                                      {...defaultSideNavProps}/>)}
-            {this.renderRoute(IslandRoutes.RansomwareReport,
-              <SidebarLayoutComponent component={ReportPage}
-                                      {...defaultSideNavProps}/>)}
-            {this.renderRoute(IslandRoutes.LicensePage,
-              <SidebarLayoutComponent component={LicensePage}
-                                      {...defaultSideNavProps}/>)}
+            <Route path={IslandRoutes.RegisterPage}
+                   element={<RegisterPageComponent onStatusChange={this.updateStatus}/>}/>
+            <Route element={<PluginsProvider value={pluginsContextValue}/>}>
+              {this.renderRoute(IslandRoutes.GettingStartedPage,
+                <SidebarLayoutComponent component={GettingStartedPage} {...defaultSideNavProps}/>)}
+              {this.renderRoute(IslandRoutes.ConfigurePage,
+                <SidebarLayoutComponent component={ConfigurePage} {...defaultSideNavProps}/>)}
+              {this.renderRoute(IslandRoutes.RunMonkeyPage,
+                <SidebarLayoutComponent component={RunMonkeyPage} {...defaultSideNavProps}/>)}
+              {this.renderRoute(IslandRoutes.MapPage,
+                <SidebarLayoutComponent component={MapPageWrapper} {...defaultSideNavProps}/>)}
+              {this.renderRoute(IslandRoutes.EventPage,
+                <SidebarLayoutComponent component={EventPage} {...defaultSideNavProps}/>)}
+              {this.renderRoute(IslandRoutes.Marketplace,
+                <SidebarLayoutComponent component={MarketplacePage} {...defaultSideNavProps}/>)}
+              {this.redirectToReport()}
+              {this.renderRoute(IslandRoutes.SecurityReport,
+                <SidebarLayoutComponent component={ReportPage}
+                                        {...defaultSideNavProps}/>)}
+              {this.renderRoute(IslandRoutes.RansomwareReport,
+                <SidebarLayoutComponent component={ReportPage}
+                                        {...defaultSideNavProps}/>)}
+              {this.renderRoute(IslandRoutes.LicensePage,
+                <SidebarLayoutComponent component={LicensePage}
+                                        {...defaultSideNavProps}/>)}
+            </Route>
             <Route path='*' element={<NotFoundPage/>}/>
           </Routes>
         </Container>
