+++ monkey/monkey_island/cc/ui/src/components/contexts/plugins/PluginsContext.js
@@ -1,0 +1,3 @@
+import React, {createContext} from 'react';
+
+export const PluginsContext = createContext({});
