+++ monkey/monkey_island/cc/ui/src/components/contexts/plugins/PluginsProvider.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import {PluginsContext} from './PluginsContext';
+import { Outlet } from 'react-router-dom';
+
+const PluginsProvider = (props) => {
+  const {value} = {...props};
+
+  return (
+    <PluginsContext.Provider value={value}>
+      <Outlet />
+    </PluginsContext.Provider>
+  )
+}
+
+export default PluginsProvider;
