--- monkey/monkey_island/cc/ui/src/components/ui-components/plugins-marketplace/PluginsContext.tsx
@@ -1,10 +1,0 @@
-import React, {createContext} from 'react';
-
-type PluginsContextType = {
-    availablePlugins: any;
-    installedPlugins: any;
-    refreshAvailablePlugins: (force: boolean) => void;
-    refreshInstalledPlugins: () => void;
-}
-
-export const PluginsContext = createContext<Partial<PluginsContextType>>({});
