+++ monkey/monkey_island/cc/ui/src/styles/components/SideNav.scss
@@ -24,4 +24,12 @@
   .general-nav-items {
     padding-left: 0;
   }
+
+  #plugins-marketplace-link .MuiBadge-badge {
+    top: 5px;
+  }
+
+  #plugins-marketplace-label {
+    margin-right: 16px;
+  }
 }
