+++ monkey/monkey_island/cc/ui/src/styles/pages/Marketplace.module.scss
@@ -1,0 +1,7 @@
+#marketplace-page {
+  :global {
+    #installed-plugins-tab-label {
+      margin-right: 16px;
+    }
+  }
+}
